renamed Namespaces to folder name
