+++ dotnet-conf-2019/DotNetConf2019.GraphQL/Data/BlogDbContext.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using DotNetConf2019.GraphQL.Models;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 
 namespace DotNetConf2019.GraphQL.Data
