+++ dotnet-conf-2019/DotNetConf2019.GraphQL/DotNetConf2019.GraphQL.csproj
@@ -14,7 +14,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Bogus" Version="28.4.4" />
     <PackageReference Include="HotChocolate.AspNetCore" Version="10.3.5" />
     <PackageReference Include="HotChocolate.AspNetCore.Playground" Version="10.3.5" />
     <PackageReference Include="HotChocolate.AspNetCore.Voyager" Version="10.3.5" />
