+++ dotnet-conf-2019/DotNetConf2019.GraphQL/Resolvers/Query.cs
@@ -1,7 +1,7 @@
 ﻿using DotNetConf2019.GraphQL.Data;
+using DotNetConf2019.GraphQL.Models;
 using HotChocolate;
 using Microsoft.EntityFrameworkCore;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -17,6 +17,9 @@ namespace DotNetConf2019.GraphQL.Schema
 
         public async Task<IReadOnlyList<Authors>> GetAuthors([Service] BlogDbContext dbContext) => 
             await dbContext.Authors.ToListAsync();
+
+        public async Task<IReadOnlyList<Comments>> GetComments([Service] BlogDbContext dbContext) =>
+            await dbContext.Comments.ToListAsync();
 
         public async Task<IReadOnlyList<Images>> GetImages([Service] BlogDbContext dbContext) =>
             await dbContext.Images.ToListAsync();
