+++ dotnet-conf-2019/DotNetConf2019.GraphQL/Types/AuthorType.cs
@@ -1,4 +1,4 @@
-﻿using DotNetConf2019.GraphQL.Data;
+﻿using DotNetConf2019.GraphQL.Models;
 using HotChocolate.Types;
 
 namespace DotNetConf2019.GraphQL.Schema
